requirements.txt
elasticsearch==9.0.2
faker==37.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0